refactor(install): generalize unpublished-snap-dependency install error
rename UpdateDependentBuildFailed -> UnpublishedComponentDependency and drop the
update-dependent wording from the message. the failing snap may no longer be tracked
in lane.updateDependents (forking a lane drops it, reset moves it out), so the message
now states the true condition: a component dependency that isn't checked out and is
pinned to a snap that was never published. detection still matches the package-manager
error against resolved component deps (the only place to recover the dep's component-id
for the bit import remediation; reads in-memory data, no fetch).
