+++ scopes/workspace/install/exceptions/index.ts
@@ -1,3 +1,3 @@
 export { DependencyTypeNotSupportedInPolicy } from './dependency-type-not-supported-in-policy';
-export { UpdateDependentBuildFailed } from './update-dependent-build-failed';
-export type { UnpublishedSnapDependency } from './update-dependent-build-failed';
+export { UnpublishedComponentDependency } from './unpublished-component-dependency';
+export type { UnpublishedSnapDependency } from './unpublished-component-dependency';
